Move the Gutter menu and editor context menu into their own "component" functions.
